Extract persona customize-page categories into PersonaClassBase
Moves getCustomizePageCategories/getCustomizePageOptions out of
PersonaUtils into an overridable PersonaClassBase, following the same
ClassBase pattern used elsewhere in the UI. Lets downstream products
extend the persona customize-page category list without forking
PersonaUtils.
